Tomcat random number fix
Added java opt to uranium as thee is a known issue discovered during
the cloud foundry roadshows where many users starting containers on a
DEA, then this is resulting in errors on the health manager as each
instance takes longer to start (as by default /dev/random) is used.
This fix was suggested by Vik Rana of the PWS team.